Renamed module files to mgp and _mgp
